Visualizer compose based ready 21.11.2023

+++ playerFeature/src/main/java/com/yes/player/presentation/model/PlayerStateUI.kt
@@ -8,5 +8,5 @@ data class PlayerStateUI(
     val durationInt:Int?=null,
     val stateBuffering:Boolean=false,
     val progress:Int?=null,
-    val visualizerData:DoubleArray?=null
+    val visualizerData:FloatArray?=null
 )
